python_app/pyproject.toml
[project]
name = "photos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "geopy>=2.4.1",
    "google-cloud-aiplatform>=1.91.0",
    "pillow>=11.2.1",
    "pinecone>=6.0.2",
    "tqdm>=4.67.1",
    "pyside6>=6.8.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
]
